+++ utilBMF/PCRSimulation.pyx
@@ -15,7 +15,7 @@ Massively Parallel Sequencing datasets.
 """
 
 
-@cython.locals(countMC=cython.int, size=cython.int)
+@cython.locals(countMC=cython.long, size=cython.long)
 @cython.boundscheck(False)
 @cython.wraparound(False)
 def main():
@@ -63,7 +63,7 @@ def main():
     return
 
 
-@cython.locals(size=cython.int, i=cython.int)
+@cython.locals(size=cython.long, i=cython.long)
 @cython.boundscheck(False)  # Turn off boundscheck for this function
 @cython.wraparound(False)  # Turn off negative indexing
 @cython.returns(DTYPE_t)
